requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
groq
sentence-transformers
transformers
qdrant-client
rank-bm25
celery
redis